requirements.txt
langchain>=0.2.0
langgraph>=0.2.0
google-genai>=1.0.0
faiss-cpu>=1.8.0
sentence-transformers>=3.0.0
datasets>=2.19.0
pandas>=2.2.0
numpy>=1.26.0
scipy>=1.13.0
scikit-learn>=1.4.0
wandb>=0.17.0
matplotlib>=3.9.0
seaborn>=0.13.0
plotly>=5.22.0
streamlit>=1.35.0
python-dotenv>=1.0.0
tqdm>=4.66.0
pydantic>=2.7.0
requests
ragas>=0.1.9
# Fine-tuning (run on Colab GPU, not locally):
# transformers>=4.41.0
# peft>=0.11.0
# bitsandbytes>=0.43.0
# trl>=0.9.0
# accelerate>=0.30.0
